Added latitude and longitude for most tour stop locations...still need 6 more
Added code in updateMap to place two markers on map and zoom to the two points

+++ app/src/main/java/com/example/joseph/tlucampustour/TourCursorAdapter.java
@@ -34,8 +34,8 @@ public class TourCursorAdapter extends CursorAdapter {
     {
         Cursor myCursor = getCursor();
         String name = myCursor.getString(NAME_COL_POSITION);
-        float lat = myCursor.getFloat(LAT_COL_POSITION);
-        float lon = myCursor.getFloat(LONG_COL_POSITION);
+        double lat = myCursor.getDouble(LAT_COL_POSITION);
+        double lon = myCursor.getDouble(LONG_COL_POSITION);
         int info = myCursor.getInt(INFO_COL_POSITION);
         int img = myCursor.getInt(IMG_COL_POSITION);
         int audio = myCursor.getInt(AUDIO_COL_POSITION);
